chore: begins major code refactor
