+++ vCardLib.Tests/AddressCollectionTests.cs
@@ -1,5 +1,8 @@
 ﻿using System;
 using NUnit.Framework;
+using vCardLib.Collections;
+using vCardLib.Models;
+
 namespace vCardLib.Tests
 {
 	[TestFixture]
