+++ vCardLib.Tests/AddressTests.cs
@@ -1,4 +1,6 @@
 ﻿using NUnit.Framework;
+using vCardLib.Models;
+
 namespace vCardLib.Tests
 {
 	[TestFixture]
