+++ vCardLib.Tests/HelperTests.cs
@@ -3,6 +3,9 @@ using System;
 using System.IO;
 using System.Reflection;
 using NUnit.Framework;
+using vCardLib.Collections;
+using vCardLib.Helpers;
+
 namespace vCardLib.Tests
 {
 	[TestFixture]
