+++ vCardLib.Tests/HobbyCollectionTests.cs
@@ -1,6 +1,9 @@
 ﻿// Created by Bolorunduro Winner-Timothy on  11/22/2016 at 5:18 AM
 using System;
 using NUnit.Framework;
+using vCardLib.Collections;
+using vCardLib.Models;
+
 namespace vCardLib.Tests
 {
 	[TestFixture]
