+++ vCardLib.Tests/InterestTests.cs
@@ -1,5 +1,6 @@
 ﻿// Created by Bolorunduro Winner-Timothy on  11/22/2016 at 5:33 AM
 using NUnit.Framework;
+using vCardLib.Models;
 
 namespace vCardLib.Tests
 {
