+++ vCardLib.Tests/PhoneNumberCollectionTests.cs
@@ -1,6 +1,8 @@
 ﻿// Created by Bolorunduro Winner-Timothy on  11/22/2016 at 5:38 AM
 using System;
 using NUnit.Framework;
+using vCardLib.Collections;
+using vCardLib.Models;
 
 namespace vCardLib.Tests
 {
