+++ vCardLib.Tests/vCardTests.cs
@@ -2,6 +2,9 @@
 using System;
 using System.IO;
 using System.Reflection;
+using vCardLib.Collections;
+using vCardLib.Helpers;
+using vCardLib.Models;
 
 namespace vCardLib.Tests
 {
