+++ vCardLib/Models/Photo.cs
@@ -3,7 +3,7 @@ using System.Drawing;
 using System.Drawing.Imaging;
 using System.IO;
 
-namespace vCardLib
+namespace vCardLib.Models
 {
     /// <summary>
     /// Supported ecnoding types
